Improve the structuring of code using __init__.py
I don't use python much. This is important however and a pretty nice
change all around considering the fact that it allows us to import in an
upward direction too instead of strictly downwards. This means modules
can import other sibling modules for example, which if we're considering
using typing is an important stepping stone to do.