Add poetry
Not the cleanest setup, but that's mainly because I was trying to make
it work with the existing directory setup. Ideally I'd use the
conventional poetry file hierarchy, but for now this works. Another
thing I encountered is that poetry run has to be done in the src/
directory because otherwise I *think* it expects you to do imports like
src.cogs.help or whatever.