Add flake.nix flake.lock
These two files are nice because that means that this can be easily used
and ran on a nix setup. The cool thing about nix is like poetry, it's a
reproducible build system that can be used on not just python projects,
but basically anything. Poetry2nix is a way to get nix to understand
poetry projects.